[WIP] use 'debug' module instead of `console.log`. split out more functions. rename some functions.
(no idea if any of this works still/yet, and still very old code.)

also removed `checkAffiliateUrl`, seems too narrow to be in this module.
