+++ lib/parser.js
@@ -1,9 +1,13 @@
-var _ = require('lodash');
+var
+  _ = require('lodash'),
+  flatten = require('./flatten').flatten,
+  debug = require('debug')('ebay:parser');
+
 
 // extract an array of items from responses. differs by query type.
 // @todo build this out as more queries are added...
 exports.parseItemsFromResponse = function parseItemsFromResponse(data, callback) {
-  // console.log('parse data', data);
+  debug('parsing items', data);
 
   var items = [];
   try {
