+++ lib/urls.js
@@ -1,4 +1,7 @@
-var _ = require('lodash');
+var
+  _ = require('lodash'),
+  debug = require('debug')('ebay:urls');
+
 
 // [internal] convert params hash to url string.
 // some items may be arrays, use key(0)..(n)
@@ -32,10 +35,10 @@ function buildUrlParams(obj, prefix) {
     }
   } else {
     str.push(prefix + "=" + obj);
-    console.log( prefix + " = " + obj);
+    debug(prefix + " = " + obj);
   }
   return str.join("&");
-};
+}
 
 
 // [internal] convert a filters array to a url string
@@ -61,7 +64,7 @@ function buildFilters(filterType, filters) {
     });
   });
   return urlFilter;
-};
+}
 
 
 // build URL to API endpoints
